Reconstruct tracer package
